+++ include/Matrix.hpp
@@ -13,29 +13,29 @@
 class Matrix{
     private:
         std::vector<float> matrixValue;
-        int row_qtity;
-        int col_qtity;
+        size_t row_qtity;
+        size_t col_qtity;
 
 
     public:
-        Matrix(int succ_layer_dim, int prec_layer_dim): matrixValue(succ_layer_dim*prec_layer_dim, rand()% RAND_M), row_qtity(prec_layer_dim) , col_qtity(succ_layer_dim) {}
+        Matrix(size_t succ_layer_dim, size_t prec_layer_dim): matrixValue(succ_layer_dim*prec_layer_dim, rand()% RAND_M), row_qtity(prec_layer_dim) , col_qtity(succ_layer_dim) {}
         
         // initialize an array of dimention 1 on the fly
         Matrix(std::vector<float> values);
 
         // returns the quantity of rows and cols
-        int getRowQuantity() const;
-        int getColQuantity() const;
+        size_t getRowQuantity() const;
+        size_t getColQuantity() const;
 
         // get or put a value on the matrixValue vector
-        float getValue(int) const;
-        void putValue(int, float);
+        float getValue(size_t) const;
+        void putValue(size_t, float);
 
         // empties the matrix. Gurantees no reallocation
         void clear();
 
         // converts from 2d coordinates to array coordinates. 
-        static int convert_dimention(int row_number, int col_number, int col_quantity);
+        static size_t convert_dimention(size_t row_number, size_t col_number, size_t col_quantity);
 
 
 };
